Refactor analysis module to use f32 instead of f64 for features and embeddings
- Updated the `Feature` type to f32 across the analysis module.
- Changed data types in various structs and functions to use f32 for consistency.
- Modified tests and benchmarks to accommodate the new feature type.
- Adjusted utility functions and data processing methods to handle f32.
- Ensured compatibility with existing data by converting f64 to f32 where necessary.

+++ analysis/src/decoder/mecomp.rs
@@ -433,7 +433,7 @@ mod tests {
         );
     }
 
-    const PATH_AND_EXPECTED_ANALYSIS: (&str, [f64; NUMBER_FEATURES]) = (
+    const PATH_AND_EXPECTED_ANALYSIS: (&str, [f32; NUMBER_FEATURES]) = (
         "data/s16_mono_22_5kHz.flac",
         [
             0.384_638_9,
@@ -477,7 +477,7 @@ mod tests {
         }
     }
 
-    const RESAMPLED_PATH_AND_EXPECTED_ANALYSIS: (&str, [f64; NUMBER_FEATURES]) = (
+    const RESAMPLED_PATH_AND_EXPECTED_ANALYSIS: (&str, [f32; NUMBER_FEATURES]) = (
         "data/s32_stereo_44_1_kHz.flac",
         [
             0.38463664,
